backend/requirements.txt
fastapi==0.128.2
uvicorn==0.40.0
python-multipart
pillow
numpy
ultralytics==8.4.12
opencv-python-headless
